Enforce maximum upload size (#329)
* Expose part size from ObjectClient

Some users of the client need to know what part size it's using
internally. Rather than plumbing that around everywhere, and deferring
to CRT defaults if unspecified, let's just make it explicit everywhere.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Allow FUSE tests to set a part size

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Enforce maximum upload size

Uploads can fail if they require more than 10,000 parts. Right now we
only find out about that failure asynchronously when the CRT actually
constructs the 10,0001th part. This is bad because unlike other upload
errors, this one is a deterministic failure that we should be able to
report up front.

One trick here is that the kernel appears to retry failed writes from
page cache when returning EFBIG, so we need to remember the reason the
write failed.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Enforce part size bounds

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>