Capture panics and pipe them to `tracing`
We want to make sure panics are captured in logs so we can debug them.
This change adds a new panic hook that pulls out the payload, location,
and backtrace of a panic and emits them at error level through
`tracing`. The `main.rs` file is getting a bit out of hand so I tried to
leave it a little better than I found it by moving the logging stuff
into a submodule.

Signed-off-by: James Bornholt <bornholt@amazon.com>