Reconcile remote and existing inodes at `update` time
To date we haven't thought too carefully about what happens if objects
are put/deleted from the S3 bucket while conflicting state is present
locally. There are a lot of edge cases here -- the Cartesian product of
existing state (local/remote file/directory) and new remote state
(file/directory), as well as two paths for inodes to be updated (readdir
vs lookup).

This change defines a semantics for these permutations. The overall idea
is that (a) remote state shadows local state, and (b) directories shadow
files. But those axioms alone aren't enough to break all ties; for
example, what if the existing state is a local directory but the new
state is a remote file -- which should win? I chose to break the tie by
saying that remote directories > any local state > remote files. So, for
example, if a user creates a local directory, and then a conflicting
object appears in the remote bucket, the directory will still be
visible instead of the new file.

I spent some time trying to patch the existing inode update path to do
what I needed but it ended up being easier to just refactor it. I think
we could still find a better factoring for this path, but it now
explicitly accounts for all the permutations above and does the right
thing (at least according to our reference model) for them all.
Happily, proptest has done a good job at rooting out the many edge
cases, as you can see by all the new regression tests in this change.

Signed-off-by: James Bornholt <bornholt@amazon.com>