Remove inodes from their parent in `forget` (#380)
* Remove inodes from their parent in `forget`

The parent directory still holds onto an `Inode` (an `Arc<InodeInner>`),
so right now our `forget` is leaking the actual inode. We need to remove
it from its parent at `forget` time. Also updated the tests to check
that the inode is in fact free'd.

I tested this by listing a directory with 2M objects on an instance with
1GiB of memory, and saw constant memory usage.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Don't forget the wrong inode

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>