Implement setattr to support changing time attributes
Some applications like `touch` requires the file system to support
changing file last access and modification times. We don't support this
operation because the last modification time for objects can't be set
via S3 API. However, it's possible to allow this only for the files that
are being written because at that time it's still a temporary stat in
Mountpoint.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>