Implement setattr to support changing time attributes (#391)
* Implement setattr to support changing time attributes

Some applications like `touch` requires the file system to support
changing file last access and modification times. We don't support this
operation because the last modification time for objects can't be set
via S3 API. However, it's possible to allow this only for the files that
are being written because at that time it's still a temporary stat in
Mountpoint.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Update doc/SEMANTICS.md

Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>
Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Fix unit test

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>