Invalidate inodes when the remote object changes (#401)
* Invalidate inodes when the remote object changes

We currently try really hard to re-use inodes even when the remote
object changes. That's problematic for consistency, as the new test
added in this PR shows: if there's an existing open file handle to the
old file, the page cache can conflate the old and new file contents,
since they share an inode.

In NFS, this is solved with "generation numbers" -- we'd bump the
generation number every time the remote file changed, and the kernel
knows to invalidate file handles with outdated generations. But FUSE
didn't correctly handle generation numbers until Linux 5.13 [[1]], which
is too recent for us to rely on (e.g., AL2's kernel is 5.10).

So instead, let's just give up on reusing inodes, and enforce a new
invariant: inodes are recreated whenever the remote object changes,
including either a kind change (Directory <-> File) or an ETag change.
This lets us detect these changes and so correctly couple each file
handle to its actual object version, forbidding the kernel from sharing
caches between file versions.

The new test also exposed a bug in the prefetcher: when a request fails,
the prefetcher doesn't reset all its state, and so a subsequent read
might be to a confused/wrong offset.

[1]: https://patchwork.kernel.org/project/linux-fsdevel/patch/20210609181158.479781-1-amir73il@gmail.com/

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Move staleness check into inode

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Fix test that was assuming inodes don't change

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>