Add flags for logging configuration
This simplifies how we ask customers to configure logging by no longer
exposing them to RUST_LOG or filtering directives. Instead, we add a
`--debug` flag to enable verbose logs, and `--debug-crt` to enable
verbose CRT logs (which are spammier). We also add a `--no-log` flag to
completely turn off logging, and `--log-metrics` to emit the summarized
performance metrics.

This change also fixes a bug in the syslog implementation with events
that come via `tracing-log` (the adapter for emitting `tracing` events
from the `log` facade). We use this adapter for adapting CRT logs into
`tracing`, since `tracing` is very picky about log events having static
metadata but CRT logs are necessarily dynamic. The `tracing-log` adapter
records metadata in some custom fields, which we weren't correctly
handling, so our log messages had the wrong `target` and a bunch of
extra fields. This change is annoying to write a test for because `log`
is global, but I tested it manually.

Signed-off-by: James Bornholt <bornholt@amazon.com>