fs: refactor to use a structured error type
Today the `fs` layer's methods all return `libc::c_int`, which means the
conversion to errno happens here. This throws away a lot of error
information, and we find ourselves adding ad-hoc calls to `error!` to
save context we think is important.

This change refactors `fs`'s methods to return a new structured `Error`
type, which still includes the errno `c_int` but also a message and an
optional source (an `anyhow::Error`). We could almost use
`anyhow::Error` directly except that we need the errno conversion and
we'd like to attach a little extra context message to the errors. We
construct this wrapper with a new `err!` macro that puts the message and
source in the right place.

This change removes the ad-hoc `error!` logging we were using in a few
places in `fs` previously. I'll follow it up with another change that
adds a new proc macro to annotate every `fs` function to automatically
print its error in failure cases, which will return these log messages
and also add all the missing ones.

Signed-off-by: James Bornholt <bornholt@amazon.com>