Add new subsession IAM role for auth integration tests
We want to be able to write tests with various permutations of IAM
credentials and policies (read only, prefix only, etc). Rather than
manually building new infrastructure for them, we're creating a single
new IAM role that tests can call AssumeRole on, using a session policy
to scope down the credentials to those they want to test.

I'll be using this in a follow-up commit to switch from HeadBucket to
ListObjects for region detection. I suspect we can also use it to get
rid of our "forbidden" bucket, but don't plan on doing that right now.

Signed-off-by: James Bornholt <bornholt@amazon.com>