Use ListObjectsV2 instead of HeadBucket for region detection
HeadBucket requires `s3:ListBucket` permissions for the root of the
bucket, but some customers scope their users' access down to only a
prefix of the bucket. This makes it impossible for them to use prefix
mounts today. Instead, we want to use ListObjects on the prefix as the
region detection mechanism.

Signed-off-by: James Bornholt <bornholt@amazon.com>