Use ListObjectsV2 instead of HeadBucket for region detection  (#413)
* Refactor error handling for S3CrtClient

There are some generic errors like region redirects and permissions that
we'd like to parse in only one place rather than all over the code. This
change moves those errors to S3RequestError and rejigs the parsing code
appropriately.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Use ListObjectsV2 instead of HeadBucket for region detection

HeadBucket requires `s3:ListBucket` permissions for the root of the
bucket, but some customers scope their users' access down to only a
prefix of the bucket. This makes it impossible for them to use prefix
mounts today. Instead, we want to use ListObjects on the prefix as the
region detection mechanism.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* PR feedback

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>