Present flexible retrieval objects with 000 permissions (#414)
* Present flexible retrieval objects with 000 permissions

Objects in the Glacier Flexible Retrieval and Glacier Deep Archive
storage classes (and their corresponding Intelligent Tiering tiers) are
not readable with GetObject without first triggering a restoration. We
don't offer the ability to do that (and it likely wouldn't make sense
for us given the latency), but these objects will still appear in the
file tree, so we give them 000 permissions and EACCES failures to make
clear they're not accessible.

It would be nice to make this work for objects that have already been
restored, which still carry the GLACIER/DEEP_ARCHIVE storage class but
also return their restore state with HeadObject. But ListObjectsV2 gives
us no way to find that out. We could probably make this work since we
know we always send a HeadObject on `open`, but it was more work than I
wanted to do right now, and this at least prevents customers getting EIO
errors on these objects.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Intern storage classes

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>