Add get_identity_document to the IMDS client
Extend the IMDS client to query the identity document of the current EC2 instance. The identity document is returned as a json response, which is parsed (using serde) to extract  the region and the instance type. The new `get_identity_document()` is tested by comparing its result to those obtained via the SDK.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>