Use AWS_REGION and IMDS to determine region
Determine the region using the following sources (in order):
  * `--region` flag,
  * `AWS_REGION` environment variables,
  * EC2 instance region (using the IMDS client),
  * default region (us-east-1).

In order to avoid querying the IMDS twice, this change also refactors the retrieval of the instance type (used in the calculation of the maximum network throughput). The new type `InstanceInfo` will lazily query the IMDS client for both value if and when required.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>