Use AWS_REGION and IMDS to determine region (#422)
* Add support for AWS_REGION to determine region

Determine the region using the following sources (in order):
  * `--region` flag,
  * `AWS_REGION` environment variables,
  * default region (us-east-1).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Add get_identity_document to the IMDS client

Extend the IMDS client to query the identity document of the current EC2 instance. The identity document is returned as a json response, which is parsed (using serde) to extract  the region and the instance type. The new `get_identity_document()` is tested by comparing its result to those obtained via the SDK.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Use AWS_REGION and IMDS to determine region

Determine the region using the following sources (in order):
  * `--region` flag,
  * `AWS_REGION` environment variables,
  * EC2 instance region (using the IMDS client),
  * default region (us-east-1).

In order to avoid querying the IMDS twice, this change also refactors the retrieval of the instance type (used in the calculation of the maximum network throughput). The new type `InstanceInfo` will lazily query the IMDS client for both value if and when required.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Move InstanceInfo into a separate module

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Refactor all region handling logic into get_region

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Remove unused methods in ImdsCrtClient and ImdsClient

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>