Build SigningConfig from AuthScheme
Use the AuthScheme obtained from the EndpointResolver to build the SigningConfig
for each request. Also extend the initializer for SigningConfig to accept the
additional parameters: service name, signing algorithm, and the use_double_uri_encode
flag.

The AuthScheme will now validate the `scheme_name` field (i.e. signing algorithm)
on parsing and store it as a `SigningAlgorithm`.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>