Add an example using Mountpoint with PyTorch
I'd like to start collecting a few examples of how to use Mountpoint for
stuff. This is the first one: using Mountpoint as a PyTorch data loader.
The goal is really just to show how to do it, and maybe say a little
about how well it works.

For now, this doesn't run in CI (need a GPU instance), will work on that
later.

Signed-off-by: James Bornholt <bornholt@amazon.com>