Stub out unimplemented FUSE operations (#460)
Right now if you try to do something that's totally unsupported, like
rename, there won't be a log entry unless you're debug logging, and even
then it will only be a fuser log entry rather than something more
specific to Mountpoint. This makes it hard for customers to know what's
happening when an operation fails, and hard for us to debug with them.
So let's stub out all the FUSE methods we haven't implemented in a way
that will log the failure as a warning, like our other "unsupported"
semantics.

Signed-off-by: James Bornholt <bornholt@amazon.com>