Build releases on CentOS 7
This gets us compatibility back to glibc 2.17. The tricky part is that
CentOS 7 by default packages a GCC that's too old to build the CRT and a
Clang that's too old to run bindgen. But they also distribute optional
packages (devtoolsets) that update these toolchains and stick them in a
separate directory. So this change adopts those, and tweaks the
environment variables on the builder to point at the newer tools.

Signed-off-by: James Bornholt <bornholt@amazon.com>