Update HistogramFn impl to log failure to record rather than panic
When writing 50GiB of data to a file, mountpoint-s3 panicked due to an unexpected "ValueOutOfRangeResizeDisabled" error.
When creating the Histogram, we limit it with an upper bound of 300000000. For time durations in microsecond precision, that's 300 seconds.

This change allows the histogram to fail only with a log message for values in excess of 300 seconds.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>