Rearrange client crate and improve its docs (#511)
* Rearrange client crate and improve its docs

The current docs page for the client crate is pretty indecipherable.
While we're not currently suggesting customers use it directly, we know
a few already are, and we'd like the docs to be legible. So this commit
moves around a bunch of the public structure of the client crate to be
consistent and intentional.

There's no semantics changes here, this is all just rearranging and
commenting stuff that wasn't commented before.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Fix typo

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>