Complete the upload on flush (#526)
Currently, Mountpoint will complete an upload in two cases:
* on `release`, that is when the last file descriptor pointing to an open file handle is closed.
  This is transparent for the caller, but does not allow for reporting the outcome of the upload,
  nor for blocking until it is completed. This means that a read-after-close may not succeed
  because the upload is still in progress.
* on `fsync`, which is blocking and can return an error to the caller, but needs to be explicitly
  invoked before closing a file.

This change implements the `flush` operation, which is invoked when a file descriptor is closed.
On `flush`, like on `fsync`, Mountpoint will complete the upload, block, and return on success or
failure. In order to support common usage patterns where it is invoked multiple times, `flush`,
unlike `fsync`, will be a no-op when invoked before any data has been written or by a different
process than the one that originally opened the file.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>