Improve ChecksummedBytes::extend and clarify data integrity guarantee (#575)
* Move ChecksummedBytes into a new module

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Improve ChecksummedBytes::extend and clarify data integrity guarantee

ChecksummedBytes maintains a data buffer and a checksum and guarantees that only validated data can be accessed. Transformations such as `split_off`, `extend`, or `slice` (introduced in this change), may trigger a validation (and return an IntegrityError on failure), or propagate existing checksum(s) if possible, allowing for later validation.

This change clarifies the data integrity guarantee in the docs and optimizes the extend method to avoid re-validation when the checksums for both slices can be combined. It also avoid a redundant buffer allocation.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Add split_off tests

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

* Make clearer that shrink_to_fit does not copy data

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>