Add instance and config metadata to user agent
We'd like to record the same metadata that AWS SDKs gather so we can
better understand how different Mountpoint features are used. This
change adds support for detecting platform and instance metadata and
including it in HTTP User-agents. We follow the SDK template for
serializing this metadata.

To make this cleaner, I moved the instance info logic into the client
crate so that all users can get this kind of user agent. The new
`UserAgent` struct supports addings arbitrary key/value pairs, and we
use that in Mountpoint to record basic configurations.

User agents are always a bit annoying to test, but I manually verified
in a few cases (caching enabled/disabled) that this change was sending
the expected headers.

Signed-off-by: James Bornholt <bornholt@amazon.com>