Support S3 Express One Zone (#642)
* Support S3 Express One Zone

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Implement support for unordered readdir

ListObjectsV2 is unordered on S3 Express. This is an issue for our
shadowing semantics in readdir, which does a merge-sort of the local and
remote keys, assuming they're both ordered.

We are going to give up on readdir being ordered on Express -- it's not
required by VFS, and the alternative would be to buffer all directory
entries in memory before returning any, which would be too expensive on
large directories. So this change introduces a new unordered variant of
the readdir implementation. It just returns directory entries verbatim
from ListObjectsV2, and then returns all otherwise-unreturned local
entries at the end of the iterator.

Doing this requires Mountpoint to know whether the bucket is a directory
bucket or not, and so I added both a command-line flag and
auto-detection for that.

Signed-off-by: James Bornholt <bornholt@amazon.com>
Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Add new unordered list implementation that shuffles the entire prefix

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Add prefixed versions of list tests

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Small update to the test

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>
Signed-off-by: James Bornholt <bornholt@amazon.com>
Co-authored-by: James Bornholt <bornholt@amazon.com>