Update PyTorch example to support single files and different models (#603)
* Update PyTorch example to support single files and different models

We'd like to expand our testing to single files rather than only sharded
representations. This change adds the ability to create and train
against that style of dataset. It also adds some other useful
configurations for Mountpoint training and for configuring a different
model to train rather than hardcoding ResNet-50.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* PR feedback

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>