Prevent build on crate verify workflow
This prevents issues where some of the crates are updated but not published yet.
Cargo tries to build the crate using the version of its dependency on crates.io, as if its about to be published.
In many cases, we want to update our crates over a few commits before later publishing each of the crates together.

Example of the issue: https://github.com/awslabs/mountpoint-s3/actions/runs/7356232845/job/20026056240?pr=684#step:5:229

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>