Replace callback in FS read with simple Result instead (#691)
* Remove fs::read callback to return simple Result instead

We're making this change primarily due to the risk of a race condition introduced.
Before this change, we reply directly to the FUSE driver before exiting the fs module code.
The risk here is that we've already replied to the driver before we drop things like the file handle guard.
Albeit small, this is a race condition and we intend to remove it to avoid any risk from it.

This race condition is suspected to be the root cause for this issue
where FUSE release fails unable to unwrap the file handle reference: https://github.com/awslabs/mountpoint-s3/issues/670

This race condition risk could have a large impact since the file handle holds a reference to prefetched data.

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>

* Remove ReadReplier trait

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>