Unmount at end of fork tests (#705)
* Unmount at end of fork tests

`Command::spawn` returns a `Child`, but dropping a `Child` doesn't
shut down the process, so we leak all these mounts every time the fork
tests run. That's annoying when you run them a lot, so this change adds
unmount calls to all the tests that should succeed.

I also took this chance to clean up the test code a little by factoring
out the "wait in a loop" logic.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Update h2 dependency for https://rustsec.org/advisories/RUSTSEC-2024-0003

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Fix non-Express tests

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>