Add a `mock-mount-s3` binary that uses the mock client backend
This lets us test the entire mount-s3 binary but without a real S3
backend attached, so we can see the entire data path locally.

Because we're building a mount-s3-like binary, we should think about the
risk of confusing it with the real binary during our release process or
elsewhere. There are two safety measures in this commit that protect
against that confusion:

1. At compile time, the `mountpoint-s3-client/mock` feature must be
   enabled for this binary to compile.
2. At run time, this binary only accepts bucket names that begin with
   `sthree-`, which is documented as an invalid prefix for real S3
   bucket names.

Signed-off-by: James Bornholt <bornholt@amazon.com>