Add a new mock client that simulates GET throughput (#723)
* Add a new mock client that simulates GET throughput

For performance testing and microbenchmarking, we'd like to be able to
separate the S3 service and the CRT datapath from our own client and
file system. This mock client can simulate a target network throughput
by rate-limiting the `get_object` stream. The goal is to be able to use
this client in place of a regular `S3CrtClient` when we want to isolate
performance questions.

Along the way, I realized we're including the mock client in our release
builds, because it's an always-on feature of the client crate. This
change therefore also does a little dependency refactoring to remove the
mock and failure clients from the non-test dependency closure. I checked
this works by seeing that the release binary is a few MBs smaller, and
that `strings mount-s3 | grep mock` no longer includes the mock client's
symbols.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Adopt mock client as an option in client benchmark

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Changelog

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Dependency fixes

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>