Update to GA release of Rust SDK
This is mostly pretty straightforward with the exception of handling S3
Express One Zone, which isn't yet supported in the Rust SDK, but the SDK
is now aware of its existence. The new SDK doesn't understand the
`sigv4-express` auth scheme that Express buckets resolve to, and falls
back to no auth at all. And we can't skip over the endpoint resolution
by leaving out the bucket name any more because the SDK now validates
that bucket is present.

Our new workaround is to insert an endpoint resolver that just forces
SigV4. We still get to use the real endpoint resolver to figure out
everything else, including the endpoint URL (so that variable can be
removed). We can remove this hack once the SDK gains Express support.

Signed-off-by: James Bornholt <bornholt@amazon.com>