Update and prune some dependencies (#731)
* Update dependencies to remove some duplicate versions

Signed-off-by: James Bornholt <jamesbornholt@gmail.com>

* Sort cargo dependencies (no actual changes)

Signed-off-by: James Bornholt <jamesbornholt@gmail.com>

* Remove fs2 dependency

It's old and unmaintained, and nix has a statvfs implementation. The only trick
is that nix makes us do the block-size calculation ourselves, but since we only
care about the ratio of free blocks, we actually don't need the block size at all.

Signed-off-by: James Bornholt <jamesbornholt@gmail.com>

* Cleanup some default dependency features

A few features we either weren't using at all, or were only using in tests
and so can remove from the release build.

Signed-off-by: James Bornholt <jamesbornholt@gmail.com>

---------

Signed-off-by: James Bornholt <jamesbornholt@gmail.com>