Use stable Rust for address sanitizer
Nightly is broken today, which blocks our CI. This is the third or
fourth time this has happened to us, so let's switch over to using the
RUSTC_BOOTSTRAP hack to use nightly features on stable Rust. This is
scoped only to the ASan makefile target, so it won't actually allow us
to use nightly features in our code, just when running the sanitizers.

Signed-off-by: James Bornholt <bornholt@amazon.com>